+++ projects/vedana/CHANGELOG.md
@@ -4,6 +4,7 @@
 * (vedana-etl) Add datamodel node to graph
 * (vedana-gradio) Add GPT-5 to model selection
 * (vedana-gradio) Fix chat completion for gpt-5
+* (vedana-gradio) Remove obsolette UI components
 
 # 2025.08.07-dev.15